LayoutTests/imported/w3c/web-platform-tests/websockets/basic-auth.any-expected.txt

CONSOLE MESSAGE: WebSocket connection to 'ws://foo:bar@localhost:49001/basic_auth' failed: Unexpected response code: 401

FAIL HTTP basic authentication should work with WebSockets assert_unreached: open should succeed Reached unreachable code
